updated package

+++ src/Helpers/BasicCrudRouteHelper.php
@@ -62,11 +62,21 @@ class BasicCrudRouteHelper
              ->name("{$this->prefix}.")
              ->controller($this->controller)
              ->group(function () use ($binding) {
-                 Route::post('index', 'index')->name('index');
-                 Route::post('/', 'store')->name('store');
-                 Route::get('/', 'list')->name('list');
-                 Route::put("{{$binding}}", 'update')->name('update');
-                 Route::delete("{{$binding}}", 'destroy')->name('destroy');
+                 if (method_exists(new $this->controller, 'index')) {
+                     Route::post('index', 'index')->name('index');
+                 }
+                 if (method_exists(new $this->controller, 'store')) {
+                     Route::post('/', 'store')->name('store');
+                 }
+                 if (method_exists(new $this->controller, 'list')) {
+                     Route::get('/', 'list')->name('list');
+                 }
+                 if (method_exists(new $this->controller, 'update')) {
+                     Route::put("{{$binding}}", 'update')->name('update');
+                 }
+                 if (method_exists(new $this->controller, 'destroy')) {
+                     Route::delete("{{$binding}}", 'destroy')->name('destroy');
+                 }
              });
     }
 }
